Rename StringEx to S7String.
This name is already somewhat used in code and gives this a better name.
