+++ S7.Net/Types/StringEx.cs
@@ -1,70 +1,15 @@
 ﻿using System;
-using System.Text;
 
 namespace S7.Net.Types
 {
-    /// <summary>
-    /// Contains the methods to convert from S7 strings to C# strings
-    ///   there are two kinds how strings a send. This one is with a pre of two bytes
-    ///   they contain the length of the string
-    /// </summary>
+    /// <inheritdoc cref="S7String"/>
+    [Obsolete("Please use S7String class")]
     public static class StringEx
     {
-        /// <summary>
-        /// Converts S7 bytes to a string
-        /// </summary>
-        /// <param name="bytes"></param>
-        /// <returns></returns>
-        public static string FromByteArray(byte[] bytes)
-        {
-            if (bytes.Length < 2)
-            {
-                throw new PlcException(ErrorCode.ReadData, "Malformed S7 String / too short");
-            }
+        /// <inheritdoc cref="S7String.FromByteArray(byte[])"/>
+        public static string FromByteArray(byte[] bytes) => S7String.FromByteArray(bytes);
 
-            int size = bytes[0];
-            int length = bytes[1];
-            if (length > size)
-            {
-                throw new PlcException(ErrorCode.ReadData, "Malformed S7 String / length larger than capacity");
-            }
-
-            try
-            {
-                return Encoding.ASCII.GetString(bytes, 2, length);
-            }
-            catch (Exception e)
-            {
-                throw new PlcException(ErrorCode.ReadData,
-                    $"Failed to parse {VarType.StringEx} from data. Following fields were read: size: '{size}', actual length: '{length}', total number of bytes (including header): '{bytes.Length}'.",
-                    e);
-            }
-            
-        }
-
-        /// <summary>
-        /// Converts a <see cref="T:string"/> to S7 string with 2-byte header.
-        /// </summary>
-        /// <param name="value">The string to convert to byte array.</param>
-        /// <param name="reservedLength">The length (in bytes) allocated in PLC for string excluding header.</param>
-        /// <returns>A <see cref="T:byte[]" /> containing the string header and string value with a maximum length of <paramref name="reservedLength"/> + 2.</returns>
-        public static byte[] ToByteArray(string value, int reservedLength)
-        {
-            if (value is null)
-            {
-                throw new ArgumentNullException(nameof(value));
-            }
-
-            if (reservedLength > byte.MaxValue) throw new ArgumentException($"The maximum string length supported is {byte.MaxValue}.");
-
-            var bytes = Encoding.ASCII.GetBytes(value);
-            if (bytes.Length > reservedLength) throw new ArgumentException($"The provided string length ({bytes.Length} is larger than the specified reserved length ({reservedLength}).");
-
-            var buffer = new byte[2 + reservedLength];
-            Array.Copy(bytes, 0, buffer, 2, bytes.Length);
-            buffer[0] = (byte)reservedLength;
-            buffer[1] = (byte)bytes.Length;
-            return buffer;
-        }
+        /// <inheritdoc cref="S7String.ToByteArray(string, int)"/>
+        public static byte[] ToByteArray(string value, int reservedLength) => S7String.ToByteArray(value, reservedLength);
     }
 }
